move tests to proper folder
