+++ src/sampling/uniform.rs
@@ -66,119 +66,3 @@ pub fn sample_vec_cbd<R: RngCore + CryptoRng>(
     }
     Ok(out)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rand::rngs::OsRng;
-    use std::collections::HashMap;
-
-    #[test]
-    fn test_cbd_variance_0_5() {
-        let mut rng = OsRng;
-        let vector_size = 10000; // Large sample for statistical analysis
-        let variance = 0.5f32;
-
-        // Sample the CBD distribution
-        let samples =
-            sample_vec_cbd(vector_size, variance, &mut rng).expect("CBD sampling should succeed");
-
-        // Count occurrences of each value
-        let mut counts = HashMap::new();
-        for &sample in &samples {
-            *counts.entry(sample).or_insert(0) += 1;
-        }
-
-        println!("CBD with variance 0.5 - Value distribution:");
-        let mut sorted_keys: Vec<_> = counts.keys().collect();
-        sorted_keys.sort();
-
-        for &value in sorted_keys {
-            let count = counts[&value];
-            let percentage = (count as f64 / vector_size as f64) * 100.0;
-            println!("  Value {value}: {count} occurrences ({percentage:.1}%)");
-        }
-
-        // Verify all values are in {-1, 0, 1}
-        for &sample in &samples {
-            assert!(
-                (-1..=1).contains(&sample),
-                "Sample {sample} is outside range [-1, 1]"
-            );
-        }
-
-        // Calculate empirical variance
-        let mean: f64 = samples.iter().map(|&x| x as f64).sum::<f64>() / samples.len() as f64;
-        let variance_empirical: f64 = samples
-            .iter()
-            .map(|&x| {
-                let diff = x as f64 - mean;
-                diff * diff
-            })
-            .sum::<f64>()
-            / samples.len() as f64;
-
-        println!("Empirical statistics:");
-        println!("  Mean: {mean:.6}");
-        println!("  Variance: {variance_empirical:.6}");
-        println!("  Expected variance: 0.5");
-
-        // Verify mean is close to 0 (centered distribution)
-        assert!(
-            mean.abs() < 0.1,
-            "Mean {mean} is not close to 0 for centered distribution"
-        );
-
-        // Verify variance is close to 0.5
-        assert!(
-            (variance_empirical - 0.5).abs() < 0.1,
-            "Empirical variance {variance_empirical} is not close to expected variance 0.5"
-        );
-
-        // Verify we only get values -1, 0, 1
-        let unique_values: std::collections::HashSet<_> = samples.into_iter().collect();
-        let expected_values: std::collections::HashSet<_> = [-1, 0, 1].iter().cloned().collect();
-
-        // Check that all observed values are in the expected set
-        for value in &unique_values {
-            assert!(
-                expected_values.contains(value),
-                "Unexpected value {value} found"
-            );
-        }
-
-        println!("✓ All samples are in {{-1, 0, 1}}");
-        println!("✓ Empirical variance matches expected variance");
-        println!("✓ Distribution is properly centered");
-    }
-
-    #[test]
-    fn test_cbd_variance_1_0() {
-        let mut rng = OsRng;
-        let vector_size = 1000;
-        let variance = 1.0f32;
-
-        let samples =
-            sample_vec_cbd(vector_size, variance, &mut rng).expect("CBD sampling should succeed");
-
-        // Count occurrences
-        let mut counts = HashMap::new();
-        for &sample in &samples {
-            *counts.entry(sample).or_insert(0) += 1;
-        }
-
-        println!("\nCBD with variance 1.0 - Value distribution:");
-        let mut sorted_keys: Vec<_> = counts.keys().collect();
-        sorted_keys.sort();
-
-        for &value in sorted_keys {
-            let count = counts[&value];
-            let percentage = (count as f64 / vector_size as f64) * 100.0;
-            println!("  Value {value}: {count} occurrences ({percentage:.1}%)");
-        }
-
-        // For variance 1.0, we expect a wider range including ±2
-        let unique_values: std::collections::HashSet<_> = samples.into_iter().collect();
-        println!("Unique values with variance 1.0: {unique_values:?}");
-    }
-}
